разделил вью модель
