+++ app/src/main/java/com/dk/newprojectsearchmovie/data/common/MovieDetailCallback.kt
@@ -4,5 +4,4 @@ import com.dk.newprojectsearchmovie.data.model.imdbMovie.ImdbMovieDetail
 
 interface MovieDetailCallback {
     fun onResponse(imdbMovieDetail: ImdbMovieDetail)
-    fun onFailure()
 }
