+++ app/src/main/java/com/dk/newprojectsearchmovie/data/common/MovieListCallback.kt
@@ -4,5 +4,4 @@ import com.dk.newprojectsearchmovie.model.imdb.ImdbMovieList
 
 interface MovieListCallback {
     fun onResponse(imdbMovieList: ImdbMovieList)
-    fun onFailure()
 }
