+++ app/src/main/java/com/dk/newprojectsearchmovie/presentation/view/movielist/MainMovieListFragment.kt
@@ -61,8 +61,11 @@ class MainMovieListFragment : Fragment() {
     private fun initStorageViewModel() {
         movieListViewModel.getLocalStorage().observe(viewLifecycleOwner) {
 
-            movieListViewModel.getRequestMovieListState(MovieListType.TOP250)
-            movieListViewModel.getRequestMovieListState(MovieListType.POPULAR)
+            movieListViewModel.getRequestMovieListState(MovieListType.TOP250, movieListViewModel.getLocalStorage().value)
+            movieListViewModel.getRequestMovieListState(
+                MovieListType.POPULAR,
+                movieListViewModel.getLocalStorage().value
+            )
 
             if (it) {
                 binding.fabToggle.backgroundTintList = ColorStateList.valueOf(Color.GREEN)
